Migrate camera management code to camera session
